Major refactor: Introduce ODP in new structure (service,api,repository)
