+++ server/api/odp/createDraft.ts
@@ -1,9 +1,9 @@
 import { Express, Response, Request } from 'express'
 import { ApiAuthMiddleware } from '@server/api/middleware'
 import * as db from '@server/db/db'
-import * as odpRepository from '@server/repository/odp/odpRepository'
 import { Requests } from '@server/utils'
 import { ApiEndPoint } from '@common/api/endpoint'
+import { OdpService } from '@server/service'
 
 export const OdpCreateDraft = {
   init: (express: Express): void => {
@@ -13,7 +13,7 @@ export const OdpCreateDraft = {
       async (req: Request, res: Response) => {
         try {
           const { countryIso } = req.query
-          const result = await db.transaction(odpRepository.saveDraft, [countryIso, req.user, req.body])
+          const result = await db.transaction(OdpService.saveDraft, [countryIso, req.user, req.body])
           res.json(result)
         } catch (err) {
           Requests.sendErr(res, err)
