+++ server/api/odp/deleteDraft.ts
@@ -2,7 +2,7 @@ import { Express, Response, Request } from 'express'
 import { ApiAuthMiddleware } from '@server/api/middleware'
 import { allowedToEditDataCheck } from '@server/assessment/assessmentEditAccessControl'
 import * as db from '@server/db/db'
-import * as odpRepository from '@server/repository/odp/odpRepository'
+import { OdpService } from '@server/service'
 import { Requests } from '@server/utils'
 import { ApiEndPoint } from '@common/api/endpoint'
 
@@ -16,8 +16,8 @@ export const OdpDeleteDraft = {
           const { countryIso } = req.query
           await allowedToEditDataCheck(countryIso, req.user, 'extentOfForest')
 
-          const { odpId } = await db.transaction(odpRepository.deleteDraft, [req.query.odpId, req.user])
-          const odp = await odpRepository.getOdp(odpId)
+          const { odpId } = await db.transaction(OdpService.deleteDraft, [req.query.odpId, req.user])
+          const odp = await OdpService.getOdp(odpId)
 
           res.json({ odp })
         } catch (err) {
