+++ server/api/odp/getMany.ts
@@ -1,9 +1,9 @@
 import { Express, Response, Request } from 'express'
 import * as VersionService from '@server/service/versioning/service'
 import * as R from 'ramda'
-import * as odpRepository from '@server/repository/odp/odpRepository'
 import { Requests } from '@server/utils'
 import { ApiEndPoint } from '@common/api/endpoint'
+import { OdpService } from '@server/service'
 
 export const OdpGetMany = {
   init: (express: Express): void => {
@@ -11,9 +11,9 @@ export const OdpGetMany = {
       try {
         const schemaName = await VersionService.getDatabaseSchema(req)
 
-        const odp = R.isNil(req.query.odpId) ? Promise.resolve({}) : odpRepository.getOdp(req.query.odpId, schemaName)
+        const odp = R.isNil(req.query.odpId) ? Promise.resolve({}) : OdpService.getOdp(req.query.odpId, schemaName)
 
-        const odps = odpRepository.listOriginalDataPoints(req.query.countryIso, schemaName)
+        const odps = OdpService.listOriginalDataPoints(req.query.countryIso, schemaName)
 
         const [odpResult, odpsResult] = await Promise.all([odp, odps])
