+++ server/repository/index.ts
@@ -1,2 +1,3 @@
 export { CountryRepository } from './country'
 export { DataTableRepository } from './dataTable'
+export { OdpRepository } from './odp'
