+++ server/service/index.ts
@@ -1,2 +1,3 @@
 export { DataTableService } from './dataTable'
 export { CountryService } from './country'
+export { OdpService } from './odp'
