+++ server/service/odp/createOdp.ts
@@ -1,0 +1,3 @@
+import { OdpRepository } from '@server/repository/odp'
+
+export const { createOdp } = OdpRepository
