+++ server/service/odp/deleteDraft.ts
@@ -1,0 +1,3 @@
+import { OdpRepository } from '@server/repository/odp'
+
+export const { deleteDraft } = OdpRepository
