+++ server/service/odp/deleteOdp.ts
@@ -1,0 +1,3 @@
+import { OdpRepository } from '@server/repository/odp'
+
+export const { deleteOdp } = OdpRepository
