+++ server/service/odp/readFocOdps.ts
@@ -1,0 +1,3 @@
+import { OdpRepository } from '@server/repository/odp'
+
+export const { readFocOdps } = OdpRepository
